Created map creation system.
Currently, map uses a binary like 2d array (may be converted to binary files in the future) to generate a maze-like course for users to navigate. Actual wall collisions has yet to be implemented with this new system. The "robot" object has also been parented by the "base" object, which also required some patchwork with the Block.js file. It functions, however, readability has worsen for the time being.